+++ src/bl64-aws-env.bash
@@ -4,7 +4,7 @@
 
 # shellcheck disable=SC2034
 {
-  declare BL64_AWS_VERSION='4.0.1'
+  declare BL64_AWS_VERSION='4.0.2'
 
   declare BL64_AWS_MODULE='0'