+++ src/bl64-cnt-env.bash
@@ -4,7 +4,7 @@
 
 # shellcheck disable=SC2034
 {
-  declare BL64_CNT_VERSION='3.5.1'
+  declare BL64_CNT_VERSION='3.5.2'
 
   declare BL64_CNT_MODULE='0'