+++ src/bl64-mdb-setup.bash
@@ -16,7 +16,7 @@
 #######################################
 # shellcheck disable=SC2120
 function bl64_mdb_setup() {
-  [[ -z "$BL64_VERSION" ]] && echo 'Error: bashlib64-module-core.bash must be sourced at the end' && return 21
+  [[ -z "$BL64_VERSION" ]] && echo 'Error: bashlib64-module-core.bash must be the last sourced library' >&2 && return 21
   local mdb_bin="${1:-${BL64_VAR_DEFAULT}}"
 
   # shellcheck disable=SC2034
@@ -25,9 +25,6 @@ function bl64_mdb_setup() {
     bl64_dbg_lib_show_function "$@" &&
     _bl64_lib_module_is_imported 'BL64_MSG_MODULE' &&
     _bl64_mdb_set_command "$mdb_bin" &&
-    bl64_check_command "$BL64_MDB_CMD_MONGOSH" &&
-    bl64_check_command "$BL64_MDB_CMD_MONGORESTORE" &&
-    bl64_check_command "$BL64_MDB_CMD_MONGOEXPORT" &&
     _bl64_mdb_set_options &&
     _bl64_mdb_set_runtime &&
     BL64_MDB_MODULE="$BL64_VAR_ON"
@@ -50,31 +47,9 @@ function bl64_mdb_setup() {
 #######################################
 function _bl64_mdb_set_command() {
   bl64_dbg_lib_show_function "$@"
-  local mdb_bin="${1:-${BL64_VAR_DEFAULT}}"
-
-  if bl64_lib_var_is_default "$mdb_bin"; then
-    if [[ -x '/home/linuxbrew/.linuxbrew/bin/mongosh' ]]; then
-      mdb_bin='/home/linuxbrew/.linuxbrew/bin'
-    elif [[ -x '/opt/homebrew/bin/mongosh' ]]; then
-      mdb_bin='/opt/homebrew/bin'
-    elif [[ -x '/usr/local/bin/mongosh' ]]; then
-      mdb_bin='/usr/local/bin'
-    elif [[ -x '/opt/mongosh/bin/mongosh' ]]; then
-      mdb_bin='/opt/mongosh/bin'
-    elif [[ -x '/usr/bin/mongosh' ]]; then
-      mdb_bin='/usr/bin'
-    else
-      bl64_check_alert_resource_not_found 'mongo-shell'
-      return $?
-    fi
-  fi
-
-  bl64_check_directory "$mdb_bin" || return $?
-  [[ -x "${mdb_bin}/mongosh" ]] && BL64_MDB_CMD_MONGOSH="${mdb_bin}/mongosh"
-  [[ -x "${mdb_bin}/mongorestore" ]] && BL64_MDB_CMD_MONGORESTORE="${mdb_bin}/mongorestore"
-  [[ -x "${mdb_bin}/mongoexport" ]] && BL64_MDB_CMD_MONGOEXPORT="${mdb_bin}/mongoexport"
-
-  bl64_dbg_lib_show_vars 'BL64_MDB_CMD_MONGOSH'
+  BL64_MDB_CMD_MONGOSH="$(bl64_bsh_command_import 'mongosh' '/opt/mongosh/bin' "$@")" &&
+    BL64_MDB_CMD_MONGORESTORE="$(bl64_bsh_command_import 'mongorestore' '/opt/mongosh/bin' "$@")" &&
+    BL64_MDB_CMD_MONGOEXPORT="$(bl64_bsh_command_import 'mongoexport' '/opt/mongosh/bin' "$@")"
 }
 
 #######################################